rustls_platform_verifier

Function tls_config_with_provider

pub fn tls_config_with_provider(
    provider: Arc<CryptoProvider>,
) -> Result<ClientConfig, Error>

👎Deprecated since 0.4.0: use the BuilderVerifierExt instead

Attempts to construct a rustls configuration that verifies TLS certificates in the best way for the underlying OS platform, using the provided CryptoProvider.

See tls_config for further documentation.

Errors

Propagates any error returned by rustls::ConfigBuilder::with_safe_default_protocol_versions.