Struct rustls_platform_verifier::Verifier

pub struct Verifier { /* private fields */ }

A TLS certificate verifier that uses the system’s root store and WebPKI.

Implementations

impl Verifier

pub fn new() -> Self

Creates a new verifier whose certificate validation is provided by WebPKI, using root certificates provided by the platform.

A CryptoProvider must be set with set_provider/with_provider or CryptoProvider::install_default before the verifier can be used.

pub fn new_with_extra_roots( roots: impl IntoIterator<Item = TrustAnchor<'static>>, ) -> Self

Creates a new verifier whose certificate validation is provided by WebPKI, using root certificates provided by the platform and augmented by the provided extra root certificates.

impl Verifier

pub fn with_provider(self, crypto_provider: Arc<CryptoProvider>) -> Self

Chainable setter to configure the CryptoProvider for this Verifier.

This will be used instead of the rustls processs-default CryptoProvider, even if one has been installed.

pub fn set_provider(&mut self, crypto_provider: Arc<CryptoProvider>)

Configures the CryptoProvider for this Verifier.

This will be used instead of the rustls processs-default CryptoProvider, even if one has been installed.